Update the performance report
Include the following changes:
(1) Add the field "CNTY_ID" instead of "CNTY" in the
subregion_correlation.bin file, because "CNTY" is used by the
subregional model, and model run will fail if this performance report
added "CNTY" in this file.
(2) Corrected an error in mode choice summary. In previous version, only
the first three cores (out of 22 transit cores) are summarized.
(3) It was found that if the mode choice step is run seperately in the
stage mode, the matrix core names in some transit output file, such as
"URRPAcTrips_Tier2.mtx" are "Table 1" etc. The script is modified so
that if the first matrix core name is "Table", the core names will be
modified to appropriate names.